Move group div to allow :only-child rendering
By default, the bootstrap_mixin adds the div group in the nav.
When there are no children, this makes the `div.list-group`
have 2 elements -- the `a` link that is the group and an
empty `div`.  The empty div then means that :first-child
is used by css to render the element, when in fact it should
use :only-child.  By moving the rendering of the div inside
the `each` statement, it is only rendered when there are children
which allows bootstrap to correctly round all the corners
with CSS.